memperbaiki ke typoan
