+++ application/controllers/PembelianController.php
@@ -28,8 +28,8 @@ class PembelianController extends REST_Controller
         } else {
             $this->db->where('id', $id);
             $tokodita = $this->db->get_by_id('pembelian')->result();
-            $respon['status'] = true;
-            $respon['message'] = "berhasil mengambil semua data";
+            $respon['status'] = false;
+            $respon['message'] = "gagal mengambil semua data";
             $respon['data'] = $tokodita;
         }
         $this->response($respon, 200);
@@ -132,8 +132,8 @@ class PembelianController extends REST_Controller
             $respon['data'] = $delete;
             $this->response($respon, 200);
         } else {
-            $respon['status'] = true;
-            $respon['message'] = "berhasil menghapus data";
+            $respon['status'] = false;
+            $respon['message'] = "gagal menghapus data";
             $respon['data'] = $delete;
             $this->response($respon, 500);
         }
